Complete rework: Refactor project structure
Modularized whole code: Moved cURL, config and menu logic from source files into separate modules. Removed test and experimental files. Updated .gitignore for build directory.
