+++ lib/menuSystem.hpp
@@ -1,0 +1,13 @@
+#ifndef MENUSYSTEM_H
+#define MENUSYSTEM_H
+
+// Global variables declaration
+int opt;
+
+// Functions declaration
+void intro();
+void menu();
+void configError();
+void settings();
+
+#endif
